test: Scoreboard can create discrete instances
Adding a very basic test to prove out the tooling is working, and to
loosly assert we can create new Scoreboards. As the API relies on the
JS/TS class system, it's not a meaningful test. I had to start
somewhwere. A different API, such as one where Scoreboard was a pure
signleton, that it was "pre-intiialised" would need a different test. I
could refactor it later if it looks like such a top level interface is
better.

Contributes to task #1.