feat: add interface to retrieve scores
This adds the .getMatches() method to Scoreboard. It returns an empty array, in order to pass the initial test. However, it didn't pass the initial test, because the test was wrong: I'd used assert.equal when I should have used assert.deepEqual. As we've added a new method, this would at least be a minor release, so I've updated the version number in package.json. Contributes to #1.