feat: enable ending of matches
This adds a property  to the Match class, but not to the TMatch interface. The separation allows for type checking a restricted set of TMatch properties out of .getMatches(), whilst keeping the match data required for processing matches in the class definition. .getMatches() is updated to filter out ended matches. The Scoreboard also gains a map from home team name to the match they're playing in. This makes looking up matches by team name easier, but we must remember to remove the mapping. As these are private properties, I don't now how to assert against this behavior. It's quite important, so it should have tests, but from the perspective of core behavior, it's just an implementation detail. I might refactor it...

Refs: #3