feat: add Scoreboard and four key API methods (#7)
* docs: record initial problem domain analysis

* test: Scoreboard can create discrete instances

Adding a very basic test to prove out the tooling is working, and to
loosly assert we can create new Scoreboards. As the API relies on the
JS/TS class system, it's not a meaningful test. I had to start
somewhwere. A different API, such as one where Scoreboard was a pure
signleton, that it was "pre-intiialised" would need a different test. I
could refactor it later if it looks like such a top level interface is
better.

Contributes to task #1.

* feat: initial Scoreboard constructor

Adds a basic, empty class, to satisfy the "should allow multiple
scoreboards" test from the previous commit. The module needs an intiial
implementation to build into. Contributes to #1.

* test: getMatches() returns [] when no matches

Starting out a TDD approach to adding to an API. We only really know that we want a list of scores, and we've not added a mechanism to add scores or matches yet, so this is really the only thing we can do at this stage whilst still following TDD. Contributes to #1.

* feat: add interface to retrieve scores

This adds the .getMatches() method to Scoreboard. It returns an empty array, in order to pass the initial test. However, it didn't pass the initial test, because the test was wrong: I'd used assert.equal when I should have used assert.deepEqual. As we've added a new method, this would at least be a minor release, so I've updated the version number in package.json. Contributes to #1.

* test: new match updates getMatches result

* feat: add startMatch() method

This adds a Match class, for easy instantiation of strctured data, as well as the .startMatch() method.

refs: #1

* chore: update version in package.json

On reflection, I could probably have saved this for when this branch is merged into 'main'. Doing it that way would save me a bit of time. Ideally, I'd also create a CI job to do it automatically when merging to main. I'll add a ticket for that.

* docs: add some definitions for user stories

* test: correct a typo in test description

Refs: #1

* test: starting multiple matches grows the scoreboard

Asserts that starting multiple matches leads to the score board growing. This test passed without any code changes - I know why, and I'm content with it happening, but is it strict TDD? Could I have done that better?

Refs: #2

* test: ending a match removes it from the Scoreboard

* feat: enable ending of matches

This adds a property  to the Match class, but not to the TMatch interface. The separation allows for type checking a restricted set of TMatch properties out of .getMatches(), whilst keeping the match data required for processing matches in the class definition. .getMatches() is updated to filter out ended matches. The Scoreboard also gains a map from home team name to the match they're playing in. This makes looking up matches by team name easier, but we must remember to remove the mapping. As these are private properties, I don't now how to assert against this behavior. It's quite important, so it should have tests, but from the perspective of core behavior, it's just an implementation detail. I might refactor it...

Refs: #3

* test: assert updateScore() allows scores to be updated

* feat: add updateScore()

Finds the match to update using the home team as a key. Updates both home and away scores to match the provided data.

Refs: #4