test(orderByScoreThenIndex): assert higher scoring first param orders earlier
This commit and the next few introduce a function to sort Match objects.
The function will be supplied as a callback to `Array.prototype.sort`.
Sort callbacks accept two arguments and return -1 if the first arg
should remain "earlier" in the sorted list, 0 if the two args are the
same, and 1 if the first argument should come "later" in the sorted
list.

This adds a failing test. The passing code will be added in another
commit.

Refs: #9