feat: ensure correct results ordering (#12)
* chore: update version to reflect the match class refactor

I should have done this in the feature branch.

* test: assert results returned in score order

* test(orderByScoreThenIndex): assert higher scoring first param orders earlier

This commit and the next few introduce a function to sort Match objects.
The function will be supplied as a callback to `Array.prototype.sort`.
Sort callbacks accept two arguments and return -1 if the first arg
should remain "earlier" in the sorted list, 0 if the two args are the
same, and 1 if the first argument should come "later" in the sorted
list.

This adds a failing test. The passing code will be added in another
commit.

Refs: #9

* feat(orderByScoreThenInde): pass the initial sort order test

* docs: add notes on ordering callback

* test(ordering): assert ordering the second match when higher scored

* feat(ordering): order the second match earlier when higher scored

* test(ordering): assert secondary order by match index

* feat(ordering): secondary order by match index

* fix: remove redundant import

I think this was added accidentally due to some 'intellisense'-style helpfulness from my IDE.

* chore: bump version number

Strictly this should be a MAJOR version bump, as the sort order output of the getMatches() function has changed. However, the current version is still in the range 0.x.y and semver allows some rule flexibility 'til it reaches v1.

* fix(ordering): ensure mock matches have an index

* docs: describe the ordering utility

* chore: refactor reused types out to discrete files