test(ordering): add and pass challenge test case (#18)
* test: add test using provided data + expectations

* fix(ordering): add sort by ordering util

It turns out that I hadn't correctly finished #9! I'd made two mistakes: forgetting to add the sort clause; misreading the provided instructions such that I got the secondary sort order requirement wrong. The requirement is actually that matches with equal total score should be sorted 'youngest first' rather than 'oldest first'.

This commit fixes the first problem: it adds the sort clause to the .getMatches() function.

* fix(ordering): ensures the ordering tests reflect requirements

* fix(ordering): ensure correct secondary ordering

As well as getting the requirements wrong, I also forgot to  the sort order indicator in the case where the scores are equal. This would have stopped sorting on the secondary condition entirely, whether it followed the requirements correctly or not. This should fix it.

Refs: #9, #17