feat: block ending matches that havn't started yet (#26)
* test: assert defined errors thrown for impoosible match ends

* feat: block attempts to end matches which havn't started, or already ended

This required refactoring the tests and changing the error message from what I anticipated when I wrote them. The reason being that when ending a match, it's removed from the 'matchesByHomeTeam' array. So an ended match looks the same as a not-started match, if we only inspect that set. I could refactor to search all the mataches in the full matches array, but that would slow the operation down.

Refs: #24

* feat: when ending a match, remove it from matchesByAwayTeam map

This is just a tidy-up. I don't think there's any work done that could suffer a problem if this isn't done, but it makes more sense and pre-empts potential future bugs.

* chore: bump minor version number