feat: add getScore method
Part of the tech discussion interview. The brief was to add the .getScore() method, which would return the current score for a given team. I added a 'happy path' test, wrote the basic function, then refactored the test to cover real cases and updated the function to pass that. After the interview, I felt the process of refactoring the tests hadn't been rigourous enough - there was still a chance the method wasn't generic and able to cope with arbitrary matches, so I updated the test to prove against multiple scores.